Issue #161: Support CalChart's new Viewer File exporter with Close
Adding a new Movement for Close and changing the current Stand to Stand and Play.

Close should just print a direction, not for how long, and Stand & Play is CalBand convention, so using that syntax.

Also, updated to allow quick gen, where we can generate PDF without having to go to the next screen.